# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "ShoPlanner",
        "contact": {},
        "version": "0.0.1"
    },
    "basePath": "/api/v1",
    "paths": {
        "/auth/login": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "login with existing user",
                "operationId": "auth-login",
                "parameters": [
                    {
                        "description": "creds",
                        "name": "opts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.Credentials"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TokenResponse"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "security": [
                    {
                        "ApiAuthKey": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "logout from session",
                "operationId": "auth-logout",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "refresh access token",
                "operationId": "auth-refresh",
                "parameters": [
                    {
                        "description": "refresh token",
                        "name": "token",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/api.TokenResponse"
                        }
                    }
                }
            }
        },
        "/product": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Creates new product",
                "operationId": "product-create",
                "parameters": [
                    {
                        "description": "product to create",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.Options"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/product/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Get product info",
                "operationId": "product-get",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Product"
                ],
                "summary": "Update existing new product",
                "operationId": "product-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "product id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "product to update",
                        "name": "product",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/product.Options"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/shopmap": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShopMap"
                ],
                "summary": "Creates new shop map",
                "operationId": "shopmap-create",
                "parameters": [
                    {
                        "description": "shop map to create",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shopmap.Options"
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/shopmap/id/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShopMap"
                ],
                "summary": "Get existing shop map by it's ID",
                "operationId": "shopmap-get-id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of shop map",
                        "name": "id",
                        "in": "query"
                    }
                ],
                "responses": {}
            },
            "put": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShopMap"
                ],
                "summary": "fully updates shop map",
                "operationId": "shopmap-update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of shop map",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "new configuration",
                        "name": "config",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/shopmap.Options"
                        }
                    }
                ],
                "responses": {}
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShopMap"
                ],
                "summary": "Deletes shop map",
                "operationId": "shopmap-delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of shop map",
                        "name": "id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/shopmap/id/{id}/reorder": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShopMap"
                ],
                "summary": "only reorder categories in given shop map",
                "operationId": "shopmap-reorder",
                "parameters": [
                    {
                        "type": "string",
                        "description": "id of shop map",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "new order of categories",
                        "name": "categories",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                ],
                "responses": {}
            }
        },
        "/shopmap/user": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ShopMap"
                ],
                "summary": "Get shop maps of current logged user",
                "operationId": "shopmap-get-current-user",
                "responses": {}
            }
        },
        "/user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "list all users",
                "operationId": "user-get-all",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "creates new user",
                "operationId": "user-register",
                "parameters": [
                    {
                        "description": "data for creating new user",
                        "name": "opts",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.CreateOptions"
                        }
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "api.RefreshRequest": {
            "type": "object",
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "api.TokenResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "auth.Credentials": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string"
                },
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "product.Options": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "forms": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "shopmap.Options": {
            "type": "object",
            "properties": {
                "categories": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "viewers_id": {
                    "type": "array",
                    "uniqueItems": true,
                    "items": {
                        "type": "object"
                    }
                }
            }
        },
        "user.CreateOptions": {
            "type": "object",
            "required": [
                "password"
            ],
            "properties": {
                "login": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Auth",
            "in": "header"
        }
    }
}